Revert "Lalith-"Directory structure created""
