+++ android/app/src/main/AndroidManifest.xml
@@ -1,8 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.notification">
-   <uses-permission android:name="android.permission.INTERNET" />
    <application
-        android:label="ExpiryDate Remainder"
+        android:label="notification"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
