--- lib/globals.dart
@@ -1,3 +1,0 @@
-library notification.globals;
-
-int Selectedindex = 1;
