+++ pubspec.yaml
@@ -35,9 +35,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   flutter_barcode_scanner: ^2.0.0
-  jumping_bottom_nav_bar :
   flutter_local_notifications:
-  url_launcher:
   workmanager:
   http:
